requirements.txt
flask
requests
pdfplumber
streamlit
google-generativeai